+++ lib/import_table/workbook.rb
@@ -1,12 +1,18 @@
+require 'import_table/cell'
 require 'import_table/sheet'
 require 'import_table/options'
+require 'import_table/row'
+require 'import_table/setting'
 
 module ImportTable
   class Workbook
+    include Cell
     include Options
     include Sheet
+    include Row
+    include Setting
 
-    attr_reader :options, :settings, :info
+    attr_reader :options, :settings, :info, :uniques
 
     # @param file [String|StringIO]:
     # @param options [Hash]:
@@ -17,7 +23,7 @@ module ImportTable
     def initialize(file, options = {})
       @options = options.slice(:extension, :csv_options, :default_sheet)
       @file    = file
-
+      @uniques = {}
       review_options
       open
     end
@@ -34,14 +40,12 @@ module ImportTable
     def read(settings = {}, &block)
       @settings = settings
       change_sheet(settings[:sheet])
-      verify_rows_settings(:read)
-      prepare_mapping if @settings.include?(:mapping)
 
-      if block
-        rows_streaming(@settings[:first_row], @settings[:last_row], &block)
-      else
-        rows(@settings[:first_row], @settings[:last_row])
-      end
+      review_settings(:read)
+
+      range = @settings[:first_row].upto(@settings[:last_row])
+
+      block ? rows_streaming(range, &block) : rows(range)
     end
 
     # Read N || 10 rows for preview
@@ -54,8 +58,9 @@ module ImportTable
       @settings = settings
       change_sheet(settings[:sheet])
       verify_rows_settings(:preview)
+      range = @settings[:first_row_preview].upto(@settings[:last_row_preview])
 
-      rows(@settings[:first_row_preview], @settings[:last_row_preview])
+      rows(range)
     end
 
     private
@@ -64,76 +69,6 @@ module ImportTable
       @workbook = Roo::Spreadsheet.open(@file, @options)
 
       info!
-    end
-
-    def rows(first_row, last_row)
-      if @settings.include?(:mapping)
-        if @settings[:mapping_type] == :hash
-          first_row.upto(last_row).map { |line| prepare_row_hash(@workbook.row(line), @settings[:mapping]) }
-        end
-      else
-        first_row.upto(last_row).map { |line| @workbook.row(line) }
-      end
-    end
-
-    def rows_streaming(first_row, last_row)
-      if @settings.include?(:mapping)
-        if @settings[:mapping_type] == :hash
-          first_row.upto(last_row).each { |line| yield prepare_row_hash(@workbook.row(line), @settings[:mapping]) }
-        end
-      else
-        first_row.upto(last_row).each { |line| yield @workbook.row(line) }
-      end
-    end
-
-    def prepare_row_hash(row, mapping)
-      mapping.transform_values { |param| prepare_cell(row, param) }
-    end
-
-    def prepare_cell(row, param)
-      return nil unless param[:column]
-
-      cell_to_type(row[param[:column]], param[:type])
-    end
-
-    def cell_to_type(value, type)
-      case type
-      when 'string'
-        value
-      when 'boolean'
-        bool value
-      when 'integer'
-        Integer(value) if value
-      when 'float'
-        Float(value)
-      when 'date'
-        Date.parse(value)
-      else
-        value
-      end
-    end
-
-    def prepare_mapping
-      @settings[:mapping_type] = :hash unless @settings.include?(:mapping_type)
-
-      @settings[:mapping].each do |_, params|
-        params[:column] = ::Roo::Utils.letter_to_number(params[:column].to_s) - 1 unless params[:column].is_a?(Integer)
-      end
-    end
-
-    def verify_rows_settings(for_method = :read)
-      first, last, last_row =
-        for_method == :read ? [:first_row, :last_row, current_last_row] : [:first_row_preview, :last_row_preview, 11]
-
-      @settings[first] = @settings[first] ? verify_max_row(@settings[first]) : verify_max_row(2)
-      @settings[last]  = @settings[last] ? verify_max_row(@settings[last]) : verify_max_row(last_row)
-    end
-
-    def bool(value)
-      return true if value&.match?(/^(true|t|yes|y|1)$/i)
-      return false if value&.empty? || value&.match?(/^(false|f|no|n|0)$/i)
-
-      nil
     end
   end
 end