+++ CHANGELOG.md
@@ -7,3 +7,8 @@
 ## [0.1.0] - 2021-05-17
 ### Changed / Added
 * Remove detect mime type
+
+## [0.1.1] - 2021-05-19
+### Changed / Added
+* update: read; convert string.
+* Add: convert Date & DateTime; uniques check